error shared object mi serve il compilatore della marta

+++ src/Partition.h
@@ -13,7 +13,7 @@ private:
 public:
   std::string name="Partition";
     std::vector<unsigned int> clust_out;
-    std::vector< std::vector<unsigned int>> C;
+    std::vector< std::vector<double>> C;
     Partition();
   ~Partition();
   void update(GS_data& gs_data, const sample::GSL_RNG& gs_engine) override;
